[s] Refactoring the 'rollDice(...)' method. [c] I found it very difficult to follow what was going on in 'rollDoce()' or am refactoring it in to what I hope will be an eraser to follow structure.
Firstly I have created a "rollDiceClicked(...)" method to be the entry point into the file when the button is clicked (at first I didn't realise "rollDice(...)" was invoked from an external source).
I have moved the 'state check' block into a seperate method to 1, save on indenting and 2, let "sendStateChanges()" (the intended last line to ge executed) be the last line of the method.

The 'rollDiceClicked()' method will always call 'rollDice()' once then determine if 'resolveDice(...)' should be called.  It is the 'resolveDice(...)' method that advances to the "Buy Cards" phase.